requirements.txt
aws-cdk.aws_iam
aws-cdk.aws_eks
aws-cdk.aws_ec2
aws-cdk.core